Struct ticktock::clock::Clock

pub struct Clock {
    pub start_ns: u64,
    // some fields omitted
}

Clock structure.

Fields

start_ns

Start time of the clock, in ns since epoch

Methods

impl Clock

fn new(tick_len_ns: u64) -> Clock

Creates a new clock.

Create a clock with a tick size of tick_len_ms, in ms.

fn restart(&mut self)

Restarts the clock.

Sets the clock start time to the current time.

fn wait_until_tick(&self) -> Option<(u32, u64)>

Waits for the next clock tick.

Will wait until the next tick and return the current tick count. If the clock has become nonsensical (for example due to do a change of the system clock into the past/future, will return None.

fn iter(&mut self) -> ClockIter

Creates a clock iterator.

The iterator will iterate forever, calling wait_until_tick on each iteration. If an error occurs, it will reset the clock using restart, then resume operatior.

Returns the current tick number on each iteration.