Struct ticktock::timer::Timer

pub struct Timer {
    // some fields omitted
}

Interval-timer

Interval timers can be used to track and act on fixed time intervals when said timers are neither in charge of delay/run decision-making nor running in a separate thread:

extern crate ticktock;
use std::time;
use std::thread;
use ticktock::timer::Timer;

fn main() {
    // create two independent timers

    let mut t1 = Timer::new(time::Duration::from_millis(50));
    let mut t2 = Timer::new(time::Duration::from_millis(150));

    let delay = time::Duration::from_millis(100);

    // for testing purpose, run 10 times only
    for _ in 1..10 {
        let now = time::Instant::now();

        if t1.has_fired(now) {
            println!("T1 has fired!");

            // notify the timer it has run for this tick or it will attempt
            // to run again.
            t1.reset(now);
        }

        // the `handle()` function works like "has_fired", but updates
        // the timer automatically
        if t2.handle(now) {
            println!("T2 has fired!");
        }

        // sleep 100 ms
        thread::sleep(delay);
    }
}

Methods

impl Timer

fn new(tick_len: Duration) -> Timer

Creates a new timer.

Create a timer with a tick length of tick_len_ms in ms.

fn new_with_start_time(tick_len: Duration, start: Instant) -> Timer

Creates a new timer with a specific start time

fn handle(&mut self, t: Instant) -> bool

Combines has_fired() and reset().

fn has_fired(&self, t: Instant) -> bool

Returns true if the timer has fired since the last time passed to reset().

t is the current time and should be passed in.

fn remaining(&self, t: Instant) -> Duration

Remaining time until the timer will fire again.

fn reset(&mut self, t: Instant) -> u32

Notify the timer it has been executed.

fn set_tick_len(&mut self, tick_len: Duration, wait: bool)

Sets the tick length

If wait is true the tick length will be changed for the tick following after the current one, which might take a long time if the tick length was set to a large value before.

Passing in wait as false causes the tick length to be changed immediately, which could result in the current tick to end instantly.

fn tick_len(&self) -> Duration

Returns the current tick length