Trait ticktock::util::FloatConv

pub trait FloatConv {
    fn as_fsecs(&self) -> f64;
    fn from_fsecs(&self, fsecs: f64) -> Self;
}

Floating point conversion for time values.

Conversions will never fail, but are likely to lose precision if they cannot accurately be represented as floating point numbers. Larger values often result in smaller sub-second accuracy; large values may even lose precision at the second or greater resolution.

Required Methods

fn as_fsecs(&self) -> f64

Convert duration to f64.

Precision-loss may occur.

fn from_fsecs(&self, fsecs: f64) -> Self

Convert f64 values to time durations