Struct ticktock::clock::Clock

pub struct Clock {
    // some fields omitted
}

Clock structure.

Methods

impl Clock

fn new(tick_len: Duration) -> Clock

Creates a new clock.

Create a clock with a tick size of tick_len_ms, in ms.

fn new_with_start_time(tick_len: Duration, start: Instant) -> Clock

Creates a new clock with a specified start time

fn synced(&self, tick_len: Duration) -> Clock

Creates a new clock with a different tick length that is synced to the original clock

fn start(&self) -> Instant

Get start time

fn wait_until_tick(&self) -> (u64, Instant)

Waits for the next clock tick.

Will wait until the next tick and return the current tick count.

fn iter(&mut self) -> ClockIter

Creates a clock iterator.

The iterator will iterate forever, calling wait_until_tick on each iteration. It will panic after about 293 years.

Returns (current tick number, absolute time) on each iteration, where absolute time is relative to a fixed offset that depends on the machine (see Instant).

fn rel_iter(&mut self) -> ClockIterRelative

Create a relative clock iterator.

Similar to iter(), but the resulting iterator will return a tuple of (current tick number, relative time), with relative time being a time::Duration from the start of the clock.