Struct ticktock::clock::Clock

pub struct Clock {
    // some fields omitted
}

Clock structure.

Methods

impl Clock

fn new(tick_len: Duration) -> Clock

Creates a new clock.

Create a clock with a tick size of tick_len_ms, in ms.

fn new_with_start_time(tick_len: Duration, start: SteadyTime) -> Clock

Creates a new clock with a specified start time

fn synced(&self, tick_len: Duration) -> Clock

Creates a new clock with a different tick length that is synced to the original clock

fn start(&self) -> SteadyTime

Get start time

fn wait_until_tick(&self) -> Option<(i64, SteadyTime)>

Waits for the next clock tick.

Will wait until the next tick and return the current tick count. Returns None on clock arithmetic overflow

fn iter(&mut self) -> ClockIter

Creates a clock iterator.

The iterator will iterate forever, calling wait_until_tick on each iteration. It will panic after about 293 years.

Returns the current tick number on each iteration.